+++ packages/kit/src/views/Staking/components/ProtocolDetails/GridItemV2.tsx
@@ -1,144 +1,14 @@
-import { useMemo } from 'react';
-
-import { StyleSheet } from 'react-native';
-
-import type { IIconButtonProps, IKeyOfIcons } from '@onekeyhq/components';
-import {
-  Alert,
-  Icon,
-  IconButton,
-  Popover,
-  SizableText,
-  XStack,
-  YStack,
-} from '@onekeyhq/components';
+import { Alert, SizableText, XStack, YStack } from '@onekeyhq/components';
 import { FormatHyperlinkText } from '@onekeyhq/kit/src/components/HyperlinkText';
-import { Token } from '@onekeyhq/kit/src/components/Token';
-import { openUrlExternal } from '@onekeyhq/shared/src/utils/openUrlUtils';
 import type {
   IEarnActionIcon,
-  IEarnIcon,
-  IEarnPopupActionIcon,
   IEarnText,
-  IEarnToken,
   IEarnTooltip,
 } from '@onekeyhq/shared/types/staking';
 
 import { EarnTooltip } from '../../pages/ProtocolDetailsV2/EarnTooltip';
 
-function PopupItemLine({
-  icon,
-  title,
-  value,
-  token,
-}: {
-  icon?: IEarnIcon;
-  token?: IEarnToken;
-  title: IEarnText;
-  value: string;
-}) {
-  return (
-    <XStack gap="$2" alignItems="center" justifyContent="space-between">
-      <XStack gap="$2" alignItems="center">
-        {icon ? (
-          <Icon
-            name={icon.icon}
-            size="$4"
-            color={icon.color || '$iconSubdued'}
-          />
-        ) : null}
-        {token ? <Token tokenImageUri={token.logoURI ?? ''} size="xs" /> : null}
-        <SizableText color={title.color} size={title?.size || '$bodyMd'}>
-          {title.text}
-        </SizableText>
-      </XStack>
-      <SizableText size="$bodyMdMedium">{value}</SizableText>
-    </XStack>
-  );
-}
-
-export function ActionPopupContent({
-  bulletList,
-  items,
-  panel,
-}: {
-  bulletList: IEarnPopupActionIcon['data']['bulletList'];
-  items: IEarnPopupActionIcon['data']['items'];
-  panel: IEarnPopupActionIcon['data']['panel'];
-}) {
-  return (
-    <YStack p="$5">
-      {items?.length ? (
-        <YStack gap="$2.5">
-          {items.map(({ icon, title, value, token }) => (
-            <PopupItemLine
-              key={title.text}
-              icon={icon}
-              token={token?.info}
-              title={title}
-              value={value}
-            />
-          ))}
-        </YStack>
-      ) : null}
-      {bulletList?.length ? (
-        <YStack pt="$2" gap="$2">
-          {bulletList.map((text, index) => (
-            <XStack key={index} gap="$2" ai="center">
-              <XStack
-                h="$1"
-                w="$1"
-                my="$1.5"
-                mx="$2"
-                borderRadius="$full"
-                bg="$iconSubdued"
-              />
-              <SizableText
-                size={text.size || '$bodySm'}
-                color={text.color || '$textSubdued'}
-              >
-                {text.text}
-              </SizableText>
-            </XStack>
-          ))}
-        </YStack>
-      ) : null}
-      {panel?.length ? (
-        <XStack
-          mt="$4"
-          py="$3"
-          borderWidth={StyleSheet.hairlineWidth}
-          borderColor="$borderSubdued"
-          borderRadius="$2"
-          justifyContent="space-between"
-          width="100%"
-        >
-          {panel.map((item, index) => (
-            <YStack
-              key={index}
-              flex={1}
-              alignItems="center"
-              justifyContent="space-between"
-            >
-              <SizableText
-                color={item.title.color || '$textSubdued'}
-                size="$bodySm"
-              >
-                {item.title.text}
-              </SizableText>
-              <SizableText
-                color={item.description?.color || '$text'}
-                size="$bodyMdMedium"
-              >
-                {item.description?.text || '-'}
-              </SizableText>
-            </YStack>
-          ))}
-        </XStack>
-      ) : null}
-    </YStack>
-  );
-}
+import { EarnActionIcon } from './EarnActionIcon';
 
 export function GridItem({
   title,
@@ -153,51 +23,6 @@ export function GridItem({
   actionIcon?: IEarnActionIcon;
   type?: 'default' | 'info' | 'alert';
 }) {
-  const actionIconButton = useMemo(() => {
-    let onPress: undefined | IIconButtonProps['onPress'];
-    let icon: IKeyOfIcons | undefined;
-    switch (actionIcon?.type) {
-      case 'link':
-        icon = 'OpenOutline';
-        onPress = () => openUrlExternal(actionIcon.data.link);
-        break;
-      case 'popup':
-        return actionIcon.data.icon ? (
-          <Popover
-            floatingPanelProps={{
-              w: 360,
-            }}
-            title={title.text}
-            renderTrigger={
-              <IconButton
-                icon={actionIcon.data.icon.icon}
-                size="small"
-                variant="tertiary"
-              />
-            }
-            renderContent={
-              <ActionPopupContent
-                bulletList={actionIcon.data.bulletList}
-                items={actionIcon.data.items}
-                panel={actionIcon.data.panel}
-              />
-            }
-            placement="top"
-          />
-        ) : null;
-      default:
-    }
-    return icon ? (
-      <IconButton
-        size="small"
-        icon={icon}
-        onPress={onPress}
-        color="$iconSubdued"
-        variant="tertiary"
-      />
-    ) : null;
-  }, [actionIcon, title.text]);
-
   if (type === 'info') {
     return (
       <Alert
@@ -246,7 +71,7 @@ export function GridItem({
             {description.text}
           </FormatHyperlinkText>
         ) : null}
-        {actionIconButton}
+        <EarnActionIcon title={title.text} actionIcon={actionIcon} />
       </XStack>
     </YStack>
   );
